refactor(api/auth): restructure api authentication architecture
BREAKING CHANGES:
- Flatten use-case directory structure, removing nested folders for improved maintainability
- Move schema definitions to shared library (@byltbasics/schemas)
- Consolidate HTTP presentation layer components

Authentication:
- Refactor auth-login use case to simpler file organization
- Update auth controllers and routes to use new structure
- Improve dependency injection container for auth components

Value Objects:
- Update email, name, and password value objects to use imported schemas
- Enhance validation patterns and error handling

Presentation Layer:
- Move controllers from presentation/http/controllers to presentation/controllers
- Add dedicated validation schema middleware
- Simplify service provider structure and DI bindings
- Update Express adapter for streamlined routing

This commit significantly improves the code organization by removing unnecessary nesting
and centralizing shared components, while maintaining clean architecture principles.
