+++ apps/express-app/src/application/use-cases/auth/index.ts
@@ -1,1 +1,1 @@
-export * from "./login";
+export * from "./auth-login";
