--- apps/express-app/src/application/use-cases/auth/login/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth-login";
-export * from "./auth-login.interface";
