--- apps/express-app/src/application/use-cases/user/create-user/index.ts
@@ -1,2 +1,0 @@
-export * from "./create-user";
-export * from "./create-user.interface";
