--- apps/express-app/src/domain/auth/schemas/index.ts
@@ -1,1 +1,0 @@
-export * from "./auth-login.schema";
