--- apps/express-app/src/domain/user/schemas/index.ts
@@ -1,4 +1,0 @@
-export { createUserSchema } from "./create-user.schema";
-export { emailAddressSchema } from "./email.schema";
-export { passwordSchema } from "./password.schema";
-export { usernameSchema } from "./username.schema";
