+++ apps/express-app/src/infrastructure/providers/response-sanitizer/response-sanitizer.ts
@@ -19,7 +19,7 @@ export class ApiResponseSanitizer implements IApiResponseSanitizer {
   private readonly _defaultSanitizationRules: ISanitizationRule[] = [
     { fieldPattern: /password/i },
     { fieldPattern: /userIds/i },
-    // { fieldPattern: /token|api[_-]?key|secret/i },
+    { fieldPattern: /token|api[_-]?key|secret/i },
     { fieldPattern: /credit[_-]?card|card[_-]?number/i },
     { fieldPattern: /ssn|social[_-]?security/i },
     { fieldPattern: /auth/i },
