+++ apps/express-app/src/presentation/controllers/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./login.controller";
+export * from "./register.controller";
