+++ apps/express-app/src/presentation/controllers/auth/login.controller.ts
@@ -1,12 +1,11 @@
 import { inject, injectable } from "inversify";
 
-import type { IApiResponseSanitizer } from "@/application/providers";
 import type { IAuthLoginUseCase } from "@/application/use-cases/auth";
 import type { IAuthLoginDto, IAuthResponseDto } from "@/domain/auth/dtos";
 
 import { TYPES } from "@/infrastructure/di-container/types";
 
-import type { IHttpRequest, IHttpResponse } from "../../helpers/interfaces";
+import type { IHttpRequest } from "../../types";
 import type { IHttpController } from "../controller.interface";
 
 export interface IAuthLoginController extends IHttpController<IAuthLoginDto, IAuthResponseDto> {}
@@ -28,7 +27,7 @@ export class AuthLoginController implements IAuthLoginController {
    *
    * @param authLoginUseCase - The use case responsible for handling user login logic.
    */
-  constructor(@inject(TYPES.AuthLoginUseCase) private authLoginUseCase: IAuthLoginUseCase, @inject(TYPES.ApiResponseSanitizer) private apiResponseSanitizer: IApiResponseSanitizer) {}
+  constructor(@inject(TYPES.AuthLoginUseCase) private authLoginUseCase: IAuthLoginUseCase) {}
 
   /**
    * Handles the HTTP request to log in a user.
@@ -36,14 +35,11 @@ export class AuthLoginController implements IAuthLoginController {
    * @param request - The HTTP request containing the user login data in the body.
    * @returns A promise that resolves to an HTTP response with the authentication tokens.
    */
-  async handle(request: IHttpRequest<IAuthLoginDto>): Promise<IHttpResponse<IAuthResponseDto>> {
+  async handle(request: IHttpRequest<IAuthLoginDto>): Promise<IAuthResponseDto> {
     const body = request.body;
 
     const authTokens = await this.authLoginUseCase.execute(body);
 
-    // Sanitize the response data
-    const sanitizedResponse = this.apiResponseSanitizer.successResponse({ data: authTokens, message: "User Logged In Successfully.", statusCode: 200 });
-
-    return sanitizedResponse;
+    return authTokens;
   }
 }
