+++ apps/express-app/src/presentation/express/middlewares/validations/schemas/auth/index.ts
@@ -1,0 +1,1 @@
+export * from "./login.schema";
