+++ apps/express-app/src/presentation/express/middlewares/validations/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from "./auth";
+export * from "./user";
