+++ apps/express-app/src/presentation/express/middlewares/validations/schemas/user/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-user.schema";
