--- apps/express-app/src/presentation/http/controllers/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth-login.controller";
-export * from "./auth-register.controller";
