+++ apps/express-app/src/presentation/service-provider/index.ts
@@ -2,7 +2,7 @@
 import "reflect-metadata";
 
 import type { IApiResponseSanitizer, ILogger } from "@/application/providers";
-import type { IAuthLoginController, IAuthRegisterController } from "@/presentation/http/controllers/auth";
+import type { IAuthLoginController, IAuthRegisterController } from "@/presentation/controllers/auth";
 
 import { container } from "@/infrastructure/di-container/container";
 import { TYPES } from "@/infrastructure/di-container/types";
