+++ apps/express-app/src/presentation/types/global.ts
@@ -1,0 +1,1 @@
+export type EmptyRecord = Record<string, unknown>;
