+++ apps/express-app/src/presentation/types/index.ts
@@ -1,2 +1,2 @@
+export * from "./global";
 export * from "./http-req";
-export * from "./http-res";
